added functionality to search users via the /users endpoint
